+++ docs/versions/change_log.md
@@ -1,5 +1,9 @@
 # Change log
 
+#### `v0.4.3`
+- Added a `QueryReplay` component to assist in replaying events across a query.
+- Added `QueryErrorHandler` convenience type.
+
 #### `v0.4.2`
 - Add `append_query` method to the framework.